requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-anthropic==0.3.1
langchain-community==0.3.13
langchain-core==0.3.28
langchain-ollama==0.2.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.6
marshmallow==3.23.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.4.5
openai==1.58.1
orjson==3.10.13
packaging==24.2
pathspec==0.12.1
pip @ file:///C:/b/abs_9btgg364vx/croot/pip_1723485053292/work
platformdirs==4.3.6
propcache==0.2.1
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.0
python-dotenv==1.0.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.1.0
sniffio==1.3.1
sqlalchemy==2.0.36
tenacity==9.0.0
tiktoken==0.8.0
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.3.0
uv==0.5.13
wheel==0.44.0
yarl==1.18.3
